build: use lib name in version macros
Use the library name in version define macros. This should prevent
collisions when using the engine alongside other libraries, as the
VERSION keyword is likely very common. In addition, create three new
macros for the major, minor and patch numbers separately. This allows
the engine to use only the part it needs, when it needs it.

Signed-off-by: Danny Holman <dholman@gymli.org>